+++ aqua/core/data_model/__init__.py
@@ -1,5 +1,6 @@
 from .coordidentifier import CoordIdentifier
 from .coordtransformer import CoordTransformer, counter_reverse_coordinate
+from .coord_utils import scan_coord
 from .datamodel import DataModel
 
 __all__ = [
@@ -7,4 +8,5 @@ __all__ = [
     "CoordTransformer",
     "DataModel",
     "counter_reverse_coordinate",
+    "scan_coord",
 ]