+++ CHANGELOG.md
@@ -14,6 +14,7 @@ ClimateDT workflow modifications:
 
 Complete list:
 - Added S3 compatibility with LUMI-O bucket (#3099)
+- Update GSV paths for lumi_install (#3120)
 - Fix treatment of empty datasets in backend (#3092)
 - Cli_checker adjustments after Backend new developments (#3081)
 - Update Data Portfolio to v2.2.0 (#3084)