+++ CHANGELOG.md
@@ -13,6 +13,7 @@ ClimateDT workflow modifications:
   The most relevant change in terms of API is that `-w, --workers` become `-w, --nworkers`.
 
 Complete list:
+- Added S3 compatibility with LUMI-O bucket (#3099)
 - Update GSV paths for lumi_install (#3120)
 - Fix treatment of empty datasets in backend (#3092)
 - Cli_checker adjustments after Backend new developments (#3081)